add index limit

+++ src/lib.rs
@@ -1,4 +1,4 @@
-//! Generate the nth prime number.
+//! Generate the nth prime number. Note: limit of n index stands at [1, 1_000_000). 
 //! # Algorithm
 //! primenumbe-rs crate's algorithm is inspired by the the optimized version of the Sieve of Eratosthenes.
 //! 
@@ -9,9 +9,9 @@
 /// 
 /// ```no_run
 /// 
-/// let n: u128 = 100;
+/// let n: u64 = 100;
 /// let result = Primenumber::nthprime(n);
-/// assert_eq!(result, 541u128);
+/// assert_eq!(result, 541u64);
 /// ```
 /// 
 
@@ -19,12 +19,12 @@
 
 
 pub struct Primenumber {
-    num : u128,
+    num : u64,
 }
 
 impl Primenumber {
     
-    pub fn nthprime(n : u128) -> u128 {
+    pub fn nthprime(n : u64) -> u64 {
 
         get_nth_prime(&Primenumber { num: n })
     }
@@ -32,16 +32,16 @@ impl Primenumber {
 
 
 
-pub fn get_nth_prime(nth: &Primenumber) -> u128{
+pub fn get_nth_prime(nth: &Primenumber) -> u64{
 
-    let mut total_prime: u128 = 0;
-    let mut size_factor: u128 = 2;
+    let mut total_prime: u64 = 0;
+    let mut size_factor: u64 = 2;
 
-    let mut s : u128 = nth.num * size_factor;
-    let mut primes : Vec<u128> = Vec::new();
+    let mut s : u64 = nth.num * size_factor;
+    let mut primes : Vec<u64> = Vec::new();
     
     
-    let  n: u128 = nth.num ;
+    let  n: u64 = nth.num ;
     
     while total_prime < n {
         primes = get_primes(s).iter().copied().collect();
@@ -57,8 +57,8 @@ pub fn get_nth_prime(nth: &Primenumber) -> u128{
     return nth_prime;
 }
 
-fn get_primes(s : u128) -> Vec<u128> {
-    let mut v: Vec<u128> = vec![1; s as usize];
+fn get_primes(s : u64) -> Vec<u64> {
+    let mut v: Vec<u64> = vec![1; s as usize];
 
     for index in 2..s {
         if v[index as usize] == 1 {
@@ -76,12 +76,12 @@ fn get_primes(s : u128) -> Vec<u128> {
 }
 
 
-fn count_prime(primes : Vec<u128>, n : u128) ->Option<u128> {
-    let mut counter: u128 = 0;
+fn count_prime(primes : Vec<u64>, n : u64) ->Option<u64> {
+    let mut counter: u64 = 0;
     for i in 2..primes.iter().count() {
         counter = counter + primes.iter().nth(i).unwrap();
         if counter == n {
-            return Some(i as u128);
+            return Some(i as u64);
         }
     }
     return None;
@@ -95,8 +95,8 @@ mod tests {
 
     #[test]
     fn it_works() {
-        let n: u128 = 100;
+        let n: u64 = 100;
         let result = Primenumber::nthprime(n);
-        assert_eq!(result, 541u128);
+        assert_eq!(result, 541u64);
     }
 }
